organizing code and adding column
