+++ src/pja_asi_12c_gr2/pipelines/split_data_for_training/__init__.py
@@ -1,0 +1,3 @@
+
+
+from .pipeline import create_pipeline
